pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tts-srt-generator"
version = "0.1.0"
description = "Batch lesson generator for TTS audio and SRT transcripts"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Developer"}
]
keywords = ["tts", "srt", "audio", "transcription", "language-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "TTS>=0.22.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "pydub>=0.25.1",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "resemblyzer>=0.1.3",
    "openai-whisper>=20231117",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
tts-srt = "main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]
